backend/foodgram/requirements.txt
Django==3.2
djangorestframework==3.12.4
djoser==2.2.0
Pillow==10.0.0
python-dotenv==1.0.0
reportlab==4.0.4
django-filter==23.2
drf-extra-fields==3.6.1
django-colorfield==0.9.0
webcolors==1.11.1
psycopg2-binary==2.9.3
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
gunicorn==20.1.0